--- api/src/main/java/com/codeforcommunity/exceptions/TeamLeaderOnlyRouteException.java
@@ -1,22 +1,0 @@
-package com.codeforcommunity.exceptions;
-
-import com.codeforcommunity.rest.FailureHandler;
-import io.vertx.ext.web.RoutingContext;
-
-public class TeamLeaderOnlyRouteException extends HandledException {
-    private int teamId;
-
-    public TeamLeaderOnlyRouteException(int teamId) {
-        super();
-        this.teamId = teamId;
-    }
-
-    public int getTeamId() {
-        return teamId;
-    }
-
-    @Override
-    public void callHandler(FailureHandler handler, RoutingContext ctx) {
-        handler.handleTeamLeaderOnlyRoute(ctx, this);
-    }
-}
